Allow defining orm filters as function and move tool_id filter to separate mixin
